% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasesdeathrate-package.R
\docType{package}
\name{diseasesdeathrate-package}
\alias{diseasesdeathrate}
\alias{diseasesdeathrate-package}
\title{diseasesdeathrate: Infectious and Parasitic Disease Death Rates in Asian Countries}
\description{
This package provides data on age-standardized death rates from infectious and parasitic diseases in Asian countries, as recorded by the World Health Organization (WHO). The dataset allows users to explore health trends specific to Asian regions, making it useful for researchers, public health analysts, and data scientists with a focus on Asia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ETC5523-2024/assignment-4-packages-and-shiny-apps-yasin27mohammad/}
}

}
\author{
\strong{Maintainer}: Mohammad Yasin \email{myas0010@studnet.monash.edu}

}
\keyword{internal}
