% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeathRate.R
\docType{package}
\name{diseasesdeathrate}
\alias{diseasesdeathrate-package}
\alias{diseasesdeathrate}
\title{Disease Death Rate Analysis Package}
\description{
This package provides functions and a dataset for analyzing death rates from infectious
and parasitic diseases across Asian countries. The package includes functions to explore
trends over time and compare between countries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ETC5523-2024/assignment-4-packages-and-shiny-apps-yasin27mohammad/}
}

}
\author{
\strong{Maintainer}: Mohammad Yasin \email{myas0010@studnet.monash.edu}

}
