% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Death Rate Analysis App}
\usage{
launch_app()
}
\description{
This function launches the Shiny app for analyzing death rates from infectious and parasitic diseases.
The app provides insights into trends and comparisons across Asian countries.
}
