% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trend_of_Deathrate.R
\name{plot_death_rate_trend}
\alias{plot_death_rate_trend}
\title{Plot Death Rate Trend for a Country}
\usage{
plot_death_rate_trend(data, country)
}
\arguments{
\item{data}{A data frame containing death rate information, with columns for "Entity" (country name),
"Code" (country code), "Year", and "death rate per 100,000 population".}

\item{country}{A character string specifying the country for which to plot the death rate trend.}
}
\value{
A ggplot2 line plot showing the death rate trend for the specified country over time.
}
\description{
This function generates a line plot showing the trend of death rates from infectious and parasitic diseases for a
specified country over time. The plot displays years on the x-axis and death rate per 100,000 population on the y-axis,
allowing users to visualize changes in death rates over the years.
}
\examples{
plot_death_rate_trend(infectious_parasitic_diseases_death_rate, "Sri Lanka")

}
